waterfall: add smoothing (#33)
Adds smoothing option to the waterfall which works by applying an
image blur before colorizing the image buffer. This may be useful
for some data distributions to help reduce artifacting across
bucket width changes at the cost of reducing the effective detail
within the waterfall and potential to introduce other artifacts.

Adds smoothed versions to the simulator to help illustrate the
differences.