WIP: atomichash: initial commit
Initial commit of a hashmap implemented with atomic pointers for
each entry instead of using mutex/rwlock.

This is useful primarily if the values have interior mutability,
for example if the values are atomics.

Provides speed benefits over standard hashmap for lookups at cost
of a higher insert cost. Unlike a standard hashmap or other
concurrent hashmap implementations, the current implementation does
not resize the hash table.