ratelimiter: fix issues with token accounting
The available token accounting had some bugs which could enable too
many tokens to be in the bucket for highly contended ratelimiters.

This change replaces the fetch_add accounting for the tick with a
compare exchange which serves to eliminate races when updating the
available token count. This removes the potential for phantom
tokens to be added to the bucket if the available tokens were
reduced while doing the two step saturating add.

This should result in more accurate ratelimiting.