metrics-v2: add ThreadLocalCounter
Adds an eventually consistent counter type that uses thread local
storage to cache add/increment of counter values. This helps reduce
the number of atomic operations that are required but comes at the
cost of additional complexity. Each thread must periodically flush
the ThreadLocalCounters for updates to be seen globally.