time: refactor to improve ergonomics
The time crate allows us to represent times in a few different
ways, with differing levels of precision and both atomic and plain
representations. The initial approach led to having many types
exposed and required the caller to understand the implementation
details if a fixed representation was necessary.

This refactoring replaces the types with newtype wrappers which
allow us to see the internal representation from the calling code.
For example, `Instant` becomes `Instant<Nanoseconds<u64>>` which
conveys much more information at the callsite. `CoarseInstant`
becomes `Instant<Seconds<u32>>`.

This allows the crate to be more ergonomic and allows for other
levels of granularity without needing as many types to represent
all possible combinations.