fix(metrics): Remove public re-exports of rustcommon-time from metrics (#124)
* fix(metrics): Remove public re-exports of rustcommon-time from metrics

These were brought in to support the macros from pelikan. However, metrics should not be publically re-exporting all of rustcommon-time. This PR moves the relevant re-exports to the hidden `export` module (which is there for this exact purpose). Beyond that, I've also gone and marked the `to_lowercase!` macro as hidden along with fixing the crate name references within the macros.

* Add `crate = $crate` to all re-exported invocations of the metric macro

* Fix test cases

* Export rustcommon_time as rustcommon_metrics::time